Fix subtle edge case
Observe a short display of 'Error loading posts' when having delay=2s, variance=0 and errprob=0.2 when initially loading the app.
The culprit was the combination of two requests to the same endpoint where the second failed and debouncing inside of
search.tsx (debouncing is there to fix another visual glitch edge case).

Some changes in this commit are defensive.