Try to further fix history handling wrt manually changing urls
The idea is that when you can't guarantee a correct history index then don't try to restore the route (still need to do the same for ui states). In any case, it feels like there only ever be a compromise solution as you cannot get the correct history index in that situation.